Move jruby to other place
