+++ History.md
@@ -11,6 +11,7 @@
 * Cleanup API by moving deprecated methods to separte class
 * Cleanup Core API - reduced to `cd`, `expand_path`, `setup_aruba` and use expand_path wherever possible (fixes #253)
 * Better isolation for environment variable manipulation - really helpful from 1.0.0 on
+* Move configuration files like `jruby.rb` to `aruba/config/`-directory
 
 ## [v0.7.4](https://github.com/cucumber/aruba/compare/v0.7.2...v0.7.4)
 * Really Fixed post install message
